refactor(scan): isolate plustek-specific code
