+++ apps/scan/backend/src/index.ts
@@ -1,12 +1,13 @@
 import { LogEventId, Logger, LogSource } from '@votingworks/logging';
-import { createClient } from '@votingworks/plustek-scanner';
+import * as plustekScanner from '@votingworks/plustek-scanner';
 import * as dotenv from 'dotenv';
 import * as dotenvExpand from 'dotenv-expand';
 import fs from 'fs';
 import { NODE_ENV, SCAN_WORKSPACE } from './globals';
 import { createInterpreter, PrecinctScannerInterpreter } from './interpret';
-import * as plustekStateMachine from './state_machine';
+import * as plustekStateMachine from './scanners/plustek/state_machine';
 import * as server from './server';
+import { PrecinctScannerStateMachine } from './types';
 import { createWorkspace, Workspace } from './util/workspace';
 
 export type { Api } from './app';
@@ -57,9 +58,9 @@ async function resolveWorkspace(): Promise<Workspace> {
 function createPrecinctScannerStateMachine(
   workspace: Workspace,
   interpreter: PrecinctScannerInterpreter
-): plustekStateMachine.PrecinctScannerStateMachine {
+): PrecinctScannerStateMachine {
   return plustekStateMachine.createPrecinctScannerStateMachine({
-    createPlustekClient: createClient,
+    createPlustekClient: plustekScanner.createClient,
     workspace,
     interpreter,
     logger,
